+++ packages/amazon-bedrock/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ai-sdk/amazon-bedrock
 
+## 4.0.0-beta.68
+
+### Patch Changes
+
+- Updated dependencies [6fc35cb]
+  - @ai-sdk/anthropic@3.0.0-beta.60
+
 ## 4.0.0-beta.67
 
 ### Patch Changes