+++ packages/black-forest-labs/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @ai-sdk/black-forest-labs
 
+## 1.0.0-beta.5
+
+### Patch Changes
+
+- b8e77ef: feat(provider/black-forest-labs): Add new provider options
+
+## 1.0.0-beta.4
+
+### Patch Changes
+
+- 666bd16: fix (provider/black-forest-labs): allow null for cost and megapixel in provider response
+
+## 1.0.0-beta.3
+
+### Patch Changes
+
+- cd3b71c: feat (provider/black-forest-labs): include cost and megapixels in metadata
+
 ## 1.0.0-beta.2
 
 ### Patch Changes